requirements.txt
python-dotenv==1.0.1
typer==0.12.1
loguru==0.7.2
moviepy==1.0.3
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-speech==2.26.0
google-cloud-storage==2.16.0
# Sphinx==7.2.6
# notebook==7.1.2
